refactor: make macros tests integration tests instead of unit tests
