+++ core/src/project/mod.rs
@@ -1204,56 +1204,3 @@ mod tests {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod macro_tests {
-    use crate::project::{ProjectMut, ProjectRead, memory::InMemoryProject};
-
-    // Have to have these in scope for ProjectRead
-    // TODO: Find a better solution (that works both inside and outside sysand_core)
-    use crate::lock::Source;
-    use crate::model::{InterchangeProjectInfoRaw, InterchangeProjectMetadataRaw};
-    use crate::project::ProjectContext;
-
-    use typed_path::Utf8UnixPath;
-
-    #[derive(ProjectRead)]
-    enum NonGenericProjectRead {
-        Variant(InMemoryProject),
-    }
-
-    #[test]
-    fn test_macro_read() {
-        let _project = NonGenericProjectRead::Variant(InMemoryProject::new());
-    }
-
-    #[derive(ProjectRead, ProjectMut)]
-    enum NonGenericProjectMut {
-        Variant(InMemoryProject),
-    }
-
-    #[test]
-    fn test_macro_mut() {
-        let _project = NonGenericProjectMut::Variant(InMemoryProject::new());
-    }
-
-    #[derive(ProjectRead)]
-    enum GenericProjectRead<SomeProject: ProjectRead> {
-        Variant(SomeProject),
-    }
-
-    #[test]
-    fn test_macro_generic_read() {
-        let _project = GenericProjectRead::<InMemoryProject>::Variant(InMemoryProject::new());
-    }
-
-    #[derive(ProjectRead, ProjectMut)]
-    enum GenericProjectMut<SomeProject: ProjectRead + ProjectMut> {
-        Variant(SomeProject),
-    }
-
-    #[test]
-    fn test_macro_generic_mut() {
-        let _project = GenericProjectMut::<InMemoryProject>::Variant(InMemoryProject::new());
-    }
-}
